+++ kwimage/structs/heatmap.py
@@ -134,12 +134,18 @@ class _HeatmapDrawMixin(object):
         Args:
             channel (int | str): index of category to visualize, or a special
                 code indicating how to visualize multiple classes.
+                Can be class_idx, class_probs, or class_energy.
 
             imgspace (bool, default=False): colorize the image after
                 warping into the image space.
 
         CommandLine:
             xdoctest -m ~/code/kwimage/kwimage/structs/heatmap.py _HeatmapDrawMixin.colorize --show
+
+        Ignore:
+            import xdev
+            from kwimage.structs.heatmap import *  # NOQA
+            globals().update(xdev.get_func_kwargs(Heatmap.colorize))
 
         Example:
             >>> # xdoctest: +REQUIRES(module:kwplot)
@@ -242,6 +248,7 @@ class _HeatmapDrawMixin(object):
                 layer[..., 3] = chan
                 layer[..., 0:3] = color
                 layers.append(layer)
+
             colormask = kwimage.overlay_alpha_layers(layers)
             colormask[..., 3] *= with_alpha
             return colormask
@@ -485,6 +492,7 @@ class _HeatmapDrawMixin(object):
             >>> kwplot.autompl()
             >>> kwplot.imshow(canvas)
 
+        Ignore:
             import xdev
             globals().update(xdev.get_func_kwargs(Heatmap.draw_on))